Make stats() a little more nimble.
Replaced the path string parsing with os.path functions to increase
portability and readability
Instead of dying, print an informative error and continue when a
filename doesn't fit the expected format.
Used dateutil to parse the date in the filename more flexibly, and
handle the date as a datetime.datetime object instead of a list of ints.